+++ app/lib/data/__init__.py
@@ -1,0 +1,3 @@
+from app.lib.data.Scotland import Scotland
+from app.lib.data.Infections import Infections
+from app.lib.data.Vaccines import Vaccines
