+++ app/scripts/scraper.py
@@ -4,7 +4,7 @@ import os, re, sys, json
 import requests
 from datetime import datetime
 from bs4 import BeautifulSoup
-from scovid19.lib.Util import get_logger, project_root
+from app.lib.Util import get_logger, project_root
 
 scraper_logger = get_logger("scraper")
