+++ app/static/icons/apple-touch-icon.png
