+++ app/static/js/table.js
