+++ app/templates/infections.html.j2
