+++ app/templates/partials/chart.html.j2
