+++ app/templates/partials/settings/location_chart.html.j2
