+++ app/templates/partials/settings/trend_chart.html.j2
