+++ app/tests/responses/infections_daily.json
