+++ app/views/__init__.py
@@ -1,0 +1,2 @@
+from .api import api  # noqa: F401
+from .pages import pages  # noqa: F401
