+++ app/views/api.py
@@ -1,0 +1,57 @@
+from flask import request, Blueprint
+from app.lib.data import Infections, Vaccines
+from app.lib.Decorators import endpoint_handler
+
+
+api = Blueprint("api", __name__)
+
+
+# Misc
+@api.route("/ping")
+def ping():
+    return "Ok"
+
+
+# Infections
+@api.route("/infections/trend")
+@endpoint_handler
+def trend():
+    infections = Infections()
+    return infections.trend(request.args)
+
+
+@api.route("/infections/breakdown")
+@endpoint_handler
+def breakdown():
+    infections = Infections()
+    return infections.breakdown()
+
+
+# Vaccines
+@api.route("/vaccines/breakdown")
+@endpoint_handler
+def percentage_vaccinated():
+    vaccines = Vaccines()
+    return vaccines.percentage_vaccinated()
+
+
+@api.route("/vaccines/council")
+@endpoint_handler
+def council_breakdown():
+    vaccines = Vaccines()
+    return vaccines.council_breakdown()
+
+
+@api.route("/vaccines/trend")
+@endpoint_handler
+def vaccine_trend():
+    vaccines = Vaccines()
+    return vaccines.vaccine_trend()
+
+
+@api.route("/prevalence")
+@endpoint_handler
+def prevalence():
+    infections = Infections()
+    limit = int(request.args["limit"]) if "limit" in request.args else -1
+    return infections.prevalence()[0:limit]
