+++ config/sqliterc
@@ -1,0 +1,2 @@
+.headers on
+.mode columns
